pyproject.toml
[tool.poetry]
name = "des-application"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.4"
simpy = "^4.1.1"
django-extensions = "^3.2.3"
werkzeug = "^3.0.2"
dj-database-url = "^2.1.0"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
celery = "^5.4.0"
redis = "^5.0.3"
numpy = "^1.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
